==> Step1_Download_Statsbomb_Data.ipynb <==
%load_ext autoreload
%autoreload 2
import os; import sys;
import tqdm
import requests
import math
import zipfile

datafolder = "../data"
statsbombzip = os.path.join(datafolder, "statsbomb-open-data.zip")
statsbombroot = os.path.join(datafolder, "statsbomb-root")
statsbombdata = os.path.join(datafolder, "statsbomb-root", "open-data-master", "data")

# Create data folder if it doesn't exist
if not os.path.exists(datafolder):
    os.mkdir(datafolder)
    print(f"Directory {datafolder} created ")

url = "https://github.com/statsbomb/open-data/archive/master.zip"
# url = "https://github.com/statsbomb/open-data/tree/master/data"

# Streaming, so we can iterate over the response.
r = requests.get(url, stream=True)

block_size = 1024 * 1024
with open(statsbombzip, "wb") as f:
    datastream = tqdm.tqdm(r.iter_content(block_size),unit="MB")
    for data in datastream:
        f.write(data)

with zipfile.ZipFile(statsbombzip, 'r') as zipObj:
    zipObj.extractall(statsbombroot)

os.listdir(statsbombdata)
