==> Step7_Extra_Build_Expected_Goals_Model.ipynb <==
%load_ext autoreload
%autoreload 2
import os; import sys; sys.path.append('../')
import pandas as pd
import numpy as np
import tqdm
import warnings
warnings.simplefilter(action='ignore', category=pd.errors.PerformanceWarning)

import socceraction.classification.features as fs
import socceraction.classification.labels as lab

## Configure file and folder names
datafolder = "../data"
spadl_h5 = os.path.join(datafolder,"spadl-statsbomb.h5")
features_h5 = os.path.join(datafolder,"features.h5")
labels_h5 = os.path.join(datafolder,"labels.h5")
predictions_h5 = os.path.join(datafolder,"predictions.h5")

# Create a train and test set of games
games = pd.read_hdf(spadl_h5,"games")
games = games[games.competition_name == "FIFA World Cup"]
traingames = games[:len(games)//2]
testgames = games[len(games)//2:]
print(len(traingames),len(testgames))

# Select shots from the data and all available info about these shots

def get_shots(games):
    shots = []
    with pd.HDFStore(spadl_h5) as spadlstore,\
         pd.HDFStore(features_h5) as featurestore:
        actiontypes = spadlstore["actiontypes"]
        for game_id in tqdm.tqdm(games.game_id,desc="selecting features"):
            ai = (spadlstore[f"actions/game_{game_id}"]
                  .merge(actiontypes,how="left")
                 )
            shot_idx = ai.type_name.str.contains("shot")
            Xi = featurestore[f"game_{game_id}"]
            shots.append(Xi[shot_idx])
    return pd.concat(shots)

train_shots = get_shots(traingames)
test_shots = get_shots(testgames)

# Decide which features to use to compute the expected goals value of the shots

xfns = [fs.actiontype_onehot,
       fs.bodypart_onehot,
       fs.goalscore,
       fs.startlocation,
       fs.movement,
       fs.space_delta,
       fs.startpolar,
       fs.team,
       fs.time,
       fs.time_delta,
       ]
nb_prev_actions = 3

f = fs.feature_column_names(xfns,nb_prev_actions)
f

# Create features-matrix X and label-vector y.
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import roc_auc_score, brier_score_loss, log_loss
from xgboost import XGBClassifier

def Xy(f,shots):
    return shots[f],shots.result_success_a0

# Logistic regression
X,y = Xy(f,train_shots)
model = LogisticRegression().fit(X,y)

X,y = Xy(f,test_shots)
pred = [p[1] for p in model.predict_proba(X)]

print("ROC AUC: %.3f" % roc_auc_score(y,pred))
print("Brier score: %.3f" % brier_score_loss(y,pred))
print("Log loss: %.3f" % log_loss(y,pred))

# XGBoost
X,y = Xy(f,train_shots)
model = XGBClassifier().fit(X,y)

X,y = Xy(f,test_shots)
pred = [p[1] for p in model.predict_proba(X)]

print("ROC AUC: %.3f" % roc_auc_score(y,pred))
print("Brier score: %.3f" % brier_score_loss(y,pred))
print("Log loss: %.3f" % log_loss(y,pred))

# Naive baseline, always predict class distribution
X,y = Xy(f,train_shots)
avgP = np.mean(y)

X,y = Xy(f,test_shots)
pred = [avgP for _i in y]

print("ROC AUC: %.3f" % roc_auc_score(y,pred))
print("Brier score: %.3f" % brier_score_loss(y,pred))
print("Log loss: %.3f" % log_loss(y,pred))
